third_party/WebKit/LayoutTests/fast/css/zoom-body-scroll.html

<!DOCTYPE HTML>
<html>
<body>
    <div style="width: 1000px; height: 1000px; position: absolute; top: 0; left: 0;"></div>
    <div style="width: 100px; height: 100px; position: absolute; top: 100px; left: 0; background: green"></div>
    <p>
        Test for <i><a href="rdar://problem/6643103">rdar://problem/6643103</a>
        Unscaled values from body.scrollHeight</i>.
    </p>
    <script src="../../resources/js-test.js"></script>
    <script>
        var floatPrecision = 0.00001;
        function checkValue(actual, expected)
        {
            if (window.eventSender) {
               shouldBeCloseTo(actual, expected, floatPrecision);
            }
        }

        if (window.testRunner) {
            testRunner.dumpAsText();
            var body = document.body;
            var documentElement = document.documentElement;
            checkValue('document.body.scrollHeight', 1000)
            checkValue('document.body.scrollWidth', 1000);

            // Scrolling right to 50.
            documentElement.scrollLeft = 50;
            checkValue('document.documentElement.scrollLeft', 50);

            // Zooming in.
            eventSender.zoomPageIn();
            checkValue('document.body.scrollHeight', 1000);
            checkValue('document.body.scrollWidth', 1000);
            checkValue('document.documentElement.scrollTop', 0);
            checkValue('document.documentElement.scrollLeft', 50);

            // Scrolling down to 100.
            documentElement.scrollTop = 100;
            checkValue('document.documentElement.scrollTop', 100);

            // Zooming back out.
            eventSender.zoomPageOut();
            checkValue('document.documentElement.scrollTop', 100);
            checkValue('document.documentElement.scrollLeft', 50);
        }
    </script>
</body>
</html>
